Создаёт модуль site-menu
